+++ frontend/capacitor.config.js
@@ -1,7 +1,7 @@
 const config = {
-  appId: "com.app.smartride",
-  appName: "smartride",
-  webDir: "dist",
-  android: { allowMixedContent: true, webContentsDebuggingEnabled: true },
+    appId: "com.app.smartride",
+    appName: "smartride",
+    webDir: "dist",
+    android: { allowMixedContent: true, webContentsDebuggingEnabled: true },
 };
 export default config;
