+++ frontend/src/components/maps/widgets/AutoFocusView.tsx
