+++ frontend/src/components/maps/widgets/ClickHandler.tsx
