+++ frontend/src/components/maps/widgets/PanelButton.tsx
