+++ frontend/src/components/maps/widgets/UserFocusView.tsx
