+++ frontend/src/components/pages/LoginPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import "@/index.css";
-import { loginUser } from "@/api/auth/login";
+import { loginUser } from "@/api/services/auth/login";
 import { useAuth } from "@/components/context/useAuth.ts";
 
 const LoginPage: React.FC = () => {
